bootloader: Support PML4 and GDT being above 4GiB
Both are allocated by EFI so we have little control on their placement.
On my darp10-b both end up way above 4GiB. Fix is to add 64-bit entries to
32-bit GDT and to copy pml4 page to low memory. Then we use those temporary
artefacts in order to enter 64-bit mode nad then reload correct gdt and pml4
once we're in 64-bit mode

+++ src/system/boot/platform/efi/arch/x86/arch_smp_32.cpp
@@ -64,7 +64,7 @@ copy_trampoline_code(uint64 trampolineCode, uint64 trampolineStack)
 
 void
 prepare_trampoline_args(uint64 trampolineCode, uint64 trampolineStack,
-	uint32 pagedir, uint64 kernelEntry, addr_t virtKernelArgs,
+	uint64 pagedir, uint64 kernelEntry, addr_t virtKernelArgs,
 	uint32 currentCpu)
 {
 	trampoline_args* args = (trampoline_args *)trampolineStack;
